requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.0
python-dotenv==1.0.0
apscheduler==3.10.4